backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.31
alembic==1.13.2
aiomysql==0.2.0
PyMySQL==1.1.1

# Redis / cache
redis==5.0.7
hiredis==2.3.2

# Celery
celery==5.4.0
flower==2.0.1          # Celery monitoring UI

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
httpx==0.27.0           # OAuth HTTP calls

# Pydantic / validation
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic[email]==2.8.2
email-validator==2.2.0

# Payment SDKs
stripe==10.3.0
paypalrestsdk==1.13.3
# Airwallex uses REST API directly (no official SDK, httpx used)

# Email
resend==2.1.0
sendgrid==6.11.0

# Storage
boto3==1.34.144         # S3
# python-magic removed — replaced with pure-Python file-header detection (cross-platform)
Pillow==10.4.0          # Image processing / thumbnails

# Utils
python-slugify==8.0.4
babel==2.15.0           # i18n / locale
pytz==2024.1
python-dateutil==2.9.0
orjson==3.10.6          # Fast JSON

# Dev
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
python-dotenv==1.0.1
